Create new file app and add app class
